Fixing iterations (and many other things)
Iterations now store iteration and iterationToken in the shadow. This is used for further recomputation.
Several minor fixes in prism (mostly null-proofing).
Optimization (or more like rationalization) of script expression execution: the expression is not executed twice if there was not delta.
Extra expression tests in common.
Query dump fixes and minor output formatting fixes.