Work around failing TestResources
When applying delta with both ADD and DELETE values, we originally
added values first and then deleted them. So if a value was in both
ADD and DELETE parts, it was in fact deleted.

Now we DELETE first, and then ADD. It is not correct neither,
but let's use it as a temporary workaround for some test failures.

Correct approach is that these deltas should be forbidden, and
diff methods should not generate them.