Fix MID-4450: Role exclusion ignores relations
The exclusion constraint ignored relations altogether, triggering
even if one of conflicting roles was assigned e.g. as approver.
To resolve this, order constraints on the exclusion
are introduced (although only partially supported now).

!!! Fixed a bug in orderConstraint processing. Missing value of
relation was interpreted as "default", although it should be correctly
interpreted as "any". Deployments should be adapted. !!!