Partially fix MID-4455: Approval form errors
Originally we checked for the presence of all object wrapper
fields; now we check only those that are present on the custom
form. It is only a partial fix, as the approval forms that add
extension items currently crash for users that have no extension
before. (Contains also slight refactoring of dynamic forms code.)