Partially fix MID-4474: skipped triggers
Triggers and validation boundaries might get skipped if the trigger
or validity scanner task is brought down (via suspend or midPoint
shutdown) while processing objects. The reason is that
lastScanTimestamp is updated even if not all objects were processed.

Here we fix that. For validity boundaries it means that some of them
will be processed twice (after next task run); but it should be
harmless. Triggers are OK, as they are deleted just after being
processed.