Add tests for user password notifier (MID-4566)
Also: Fixed the user password notifier to work with secondary deltas
as well (test316UserRecompute). Fixed the notifier to stop expecting
full success.

This solution is still not complete - we have to find out somehow
if the focus update operation(s) were carried out or not. Otherwise
we could report password change even if it was not in fact executed.