Fix showing names in 'add' approvals (MID-4512)
When approving object creation, the object has an OID but does not
exist in repository. This caused problems in task details and task
lists pages.

Repository change: now we are storing targetName for references
in the XML form of stored objects (if set by caller).

(cherry picked from commit 7d7a974)