Use 'distinct' when displaying objects (MID-3955)
The GUI uses 'distinct' search option when displaying objects, unless
disabled in the system configuration.

Important change in repo: query interpreter now optimizes usage of
'distinct'. If there are no joins, distinct is not applied even if
requested. This heuristic would need to be changed if it would be
found to be ineffective (see e.g.
QueryInterpreter2Test.test0350QuerySubtreeDistinctCount).