Optimize work allocation (round 1)
Here we optimize creation of new work buckets, to be processed
in batches (instead of one-by-one). Also, when allocating existing
buckets we select them randomly, not getting always the first one.

This helps a bit: although it does not prevent collisions on coordinator
task updates, precondition violations in WorkStateManager are greatly
avoided.

TODO:
- maintain a reasonably long list of ready buckets (for randomization
  to work well)
- delegate buckets in batches (not only create them) - this will avoid
  contention on the coordinator task even more