Fix task result handling
The result was created on demand (as UNKNOWN) too eagerly: even
if it was not present just because it was not fetched. Now we create
it only on specific occasions - when transient task is created or
(with warning) when result-less task handler is to be executed.

Fixed some places that were expecting non-null result.
Also removed the coupling of result/resultStatus in repo.