Fix RExtItem concurrency issues (MID-4710)
1. Creation of RExtItem objects is now under main retry mechanism;
we removed inner retry cycle.

2. Handling of ConstraintViolationException is now part of general
exception handling, it is no longer limited to addObject
and modifyObject methods.

3. "synchronized" keyword is temporarily removed from ExtItemDictionary
methods, so we rely on DB mechanisms to ensure consistency. This
required handling of more cases of ConstraintViolationException that
occur in these situations (for all databases). This is used to test
correct working of the class in cluster environment, where java
synchronization does not protect it any more.

TODO: re-add "synchronized" keyword before 3.8.1/3.9 release.