Fix the code after cherry-pick
Together, last 2 commits introduce nameCopy attribute on Rxxx
and indices on Rxxx.name.orig. Plus removal of obsolete "toJAXB"
methods from Rxxx classes and other minor cleanup.

This is to backport MID-4537 to fix MID-4776.

(This particular commit is to fix the cherry-pick, as it required
some earlier changes that are provided here.)