Fix misreading of schema.properties (MID-4800)
When fetching schema.properties resource bundle without the context
of our specially-configured spring ResourceBundleMessageSource it is
read as ISO 8859-1 instead of UTF-8.

This commit replaces almost all such places with the use
of LocalizationService. What remains is ResourceValidatorImpl;
however, it is not used much and - most importantly - when it is
invoked, the schema.properties are already loaded.