Stop indexing dynamic ext. properties (MID-4878)
The original need was to stop indexing sync token property.
The property is not defined in any schema (it would be probably
hard e.g. because of its dynamic nature) so the most natural solution
is to avoid indexing dynamic extension properties altogether.
If someone wants to have his extension property indexed, he has to
declare it via the schema.