Remove activiti (work in progress)
This is VERY experimental implementation. Basically we try to
re-implement Activiti functionality with CaseType/CaseWorkItemType,
without changing i.e. simplifying workflow-api and overall repo-data
organization. This is left to later stages of the migration.

The implementation is not functional yet. The repo support has some
missing pieces. (So the higher levels couldn't be tested yet.)