Fix "More than one stage-level event" errors
When two work items are completed in the same moment (in two threads),
and the following stage is to be completed automatically because of
the specified condition, two stage-level completion events are
generated.

This is something that the original design didn't expected. Here we
adapt the code to such situations. Hopefully there will be no other
consequences of such simultaneous executions.