Reduce usage of searchObjectsIterative
This call was frequently used in provisioning and in search-based
expression evaluators. Its disadvantage is that it is not nicely
cacheable. So we replaced it here by searchObjects calls. The back
side is that we lost the possibility to stop search after sufficient
entries are fetched; but we didn't use this option so far anyway.

Fixes in RepositoryCache's null-or-harmless option checking are
here as well.