Add collection of bucket op stats + scavengers
We collect information on calls to getWorkBucket, completeWorkBucket,
and releaseWorkBucket and store statistics on collisions and time spent
(overall, resolving collisions, waiting for a bucket).

Also, preliminary scavengers implementation is provided, aiming to
lower the bucket processing overhead in final stages of the task run
where mis-delegated free buckets are looked for.

See MID-5405.