Fix dropping changes in LiveSync (MID-5513)
This commit fixes handling of sync token in LiveSync tasks in the
presence of task suspension and unrecoverable (halting) errors.
Token is either set to the last change successfully processed
(for "precise token" resources) or not changed at all (for all the
others).

There are also some changes in error handling. But all of this is
yet to be resolved, see MID-5514.