Add support for more relations in authorizations
Besides owner, delegator and tenant we now support requester,
assignee, and relatedObject clauses in object specification:
first two for CaseType, the third one for CaseType and TaskType.

This is an implementation of MID-3121.

Note that the authorization evaluation added is checked only on
object (CaseType/TaskType) level. In particular,
searchContainers(CaseWorkItemType) is not affected by this work.

--
Also included some other minor changes, e.g. adding "id" attribute
to schema for some containers.