Fix delegated/escalated work item deadline setting
The deadline was (for unknown reasons) determined from the original
work item deadline, if it was defined. Now it's derived solely from
the current system time.

This commit contains also a couple of related fixes aiming to unify
computation of deadline for items that are delegated/escalated at the
same time. (To correctly process their automatic rejection.)

Also adapted TestStrings to changes in workflows.