Remove cautious ext. values collision checking
Existence of extension value to be inserted is now not checked.
Only after ConstraintViolationException we engage safe algorithm
of checking the existence before operation. This can be overridden
system-wide (in repo config) or operation-wide (in RepoModifyOptions).

Whether to fetch extension values to be deleted is driven by
experimental "enableNoFetchExtensionValuesDeletion" config parameter
(default is the old behavior) and related option in RepoModifyOptions.

All of this is related to MID-5558.