Fix extra XML empty lines (MID-5645)
The issue seems to be in the built-in version of Apache Xalan present
in Java 11 runtime. So here we explicitly included latest Xalan 2.7.2
that seems to work with this respect. However, at the price of
slight performance degradation. This is to be resolved later (MID-5661).

Some namespace-related fixes and workarounds had to be introduced
as well. This is also to be researched later (MID-5661).