Fix delete/replace deltas on index-only attrs
Here we fixed the algorithm that computes the delta for repo shadow
update during change processing in provisioning. It didn't treat
index-only attributes correctly for delta-only changes. Now these
attributes are marked as incomplete when working with old repo shadow
and then updated according to the delta. This is a resolution
of MID-5832.

We also clarified fields of Change and ResourceEventDescription classes
a bit (oldShadow -> oldRepoShadow, currentShadow ->
currentResourceObject) and some of the constructors as well. Also did
similar changes in local vars at several places. However, this should
be reworked more thoroughly (MID-5834).