Rework control flow in asynchronous updates
This is a work in progress. The goal is to make asynchronous updates
execute in the worker threads of a specified task. Towards this, we
have simplified the control flow of this operation: the control is now
held in provisioning module until the task is stopped or an error
occurs.