Add renaming Grouper groups from async updates
This commit modifies ChangeProcessor to be able to fetch current
resource object when notification-only change is encountered
(if supported by the resource).

We also enhanced Dummy connector/resource to support externally-supplied
UID values (UID_MODE_EXTERNAL) and migrated Grouper story tests to use
UID_MODE_EXTERNAL.

This is a part of MID-5932 resolution. (Not complete, though. Now we
run the clockwork for each async update, which has to be selectively
stopped.)