Add "not checking in" node state
This is to differentiate between nodes that are permanently down
and ones that currently are not checking in for a short time
(between 30 and 900 seconds, by default). Also got rid of "almost dead"
terminology, replacing that by "not checking in".

(cherry picked from commit 18b4dccc6d60ef7fc9217d978f23a79dca187143)