Add incomplete items support into the repository
Items that are not stored in XML (i.e. focus.jpegPhoto, task.result,
lookupTable.row, certificationCampaign.case, index-only extension items
and shadow attributes) are now reported by the repository as incomplete,
if they are present but not requested for retrieval.

This is a step to the resolution of MID-5772 (and maybe MID-5912).