Fix "no sync policy" treatment in LiveSync (MID-5999)
Processing of such resource objects was originally considered to be
erroneous, so live sync task stopped when they were encountered.
But, in fact, these should be simply ignored.

So we added "not applicable" to the list of accepted operation result
statuses.

This commit also contains fix for live syncing account that was
recently deleted.