Fix assignment validity determination
When processing assignments, the validity of any given assignment
depends on (besides the validity of the assignment itself) the lifecycle
state of the assignment source.

The original implementation referred to the lifecycle state of the
focus. This is not correct when considering the validity of higher-order
assignments. So we switched the consideration to assignment source.

This resolves MID-6114.

(cherry picked from commit 3887e4f2d1f33679f6d15cf37e9987840a8003e6)