repo-sql-impl-test experiment with DB cleanup instead of dirty context
This speeds up the tests significantly, but currently there are couple
of problems with it:
- H2 server is sometimes started repeatedly (port binding fails).
- Seems that some objects are missing for some tests.
This may be caused by:
- Various Spring contexts are used, this is definitely a problem,
it should be unified on BaseSQLRepoTest level to avoid context refresh.
- Some data are probably initialized during Spring setup and missing
after they are deleted. We either have to add WHERE to DELETE if these
objects are always to be preserved, or (more systemic?) to initialize
these objects after the DB cleanup.