Fix op.results for threaded LiveSync/AsyncUpdate
Operation results handling in multithreaded LiveSync was faulty because
the ConnectorInstanceConnIdImpl.sync() method tried to compute,
cleanup and summarize results that were not yet complete -- they
were filled-in asynchronously, in worker threads.

So we changed this behavior and provide worker threads with their own,
separate operation results.

This change applies to AsyncUpdate as well, although the original
mechanism was a bit different there.