Fix remote access to reports (MID-6079)
Report outputs on remote nodes were not accessible because the REST
method ClusterRestService.getReportFile did not match the request.
(The reason was a mismatch in @Produces vs. Accept header.) This is now
fixed.

Moreover we added the functionality to query all cluster nodes for
report output file in situations when the file cannot be found on the
node where it is expected. This helps in situations when nodes get
renamed but the file remains stored on one of them. (MID-6079)