Add test and workaround for MID-6122
When modifying primary delta we sometimes need to clean up lens
and focus context appropriately. (Normally this should be ensured
by rot() method but it is not.)

So here we provide a test for MID-6122, showing appropriate workaround.

See MID-6132 as well.

Here we also include a small fix for tracing script variables/results.