Fix repo values for RTaskWaitingReason
In 3.9 this enum has values of 0, 1, 2. In 4.0 the "1" was removed,
shifting value of 2 to 1. Fortunately, since 4.0 we don't use values
other than 0 (at least it seems so); but in order to avoid migration
issues we now restored the deleted value of 1 as a placeholder.

Mapping of enums should be changed to be less fragile. Until that,
we CANNOT delete intermediary values from R-enums to avoid this kind
of problems.

This resolves MID-6117.