Stop visiting embedded objects by default
Visitor in prism module did NOT visit objects embedded in references
for a very long time. Recently we changed this behavior and started
visiting embedded objects - in order to assure encryption of protected
values in such objects (see 60328c40b2b99c6cf41ab6ce90145fae941d07bd).

But this change broke other things, like resolution of references
in ApprovalSchemaExecutionInformationType. Analysis of other uses
indicate that visiting embedded objects can be wrong in other
contexts as well.

So now we are reverting to the old behavior and providing
ConfigurableVisitor to allow controlling this feature.

This resolves MID-6171.