Fix CacheCounterManager memleak (MID-6343)
This is a TEMPORARY fix for memory leak in CacheCounterManager.
It is a very partial cherry-pick of d327aade9d84880a5f1f904591fd1296998ffd0e
(most of the fixes are already present in this branch).

The only thing fixed here is that counters are cleaned up for
all tasks (not only for lightweight partitioning ones). However,
because of the other changes, the leak in 4.1 was rather small.