Fix CacheCounterManager memleak (MID-6343)
This is a TEMPORARY fix for memory leak in CacheCounterManager.

1) Entries in counter manager are now created only if suspendTask
action is present.

2) Entries are cleaned up for all tasks (not only for lightweight
partitioning ones).

3) Storage of whole tasks is avoided. Only name and OID is recorded.

However, the whole fix has to be reviewed thoroughly and carried
over to master. (The state in master is already partially OK;
but not completely.)

Note that thresholds tests are failing. But they were failing
even before this commit.