Remove loading work items and cases (MID-6344)
The loadWorkItems() and loadMyRequests() methods on PageDashboard
were doing unconstrained searches for work items and cases, leading
to consumption of huge amounts (hundreds of MBs) of heap.

Output of the methods was not used anyway, so here we simple
deleted them. However, the heap consumption issue still needs to
be investigated and fixed, if necessary.