SqlRepositoryBeanConfig: explicit bean dependencies as arguments
DataSourceFactory.createDataSource() is called directly for dsf bean,
not later in sessionFactory(), otherwise it was possible to obtain DSF
without initialized datasource, which is not helpful.
This wasn't problem because ORM is initialized too, but if for whatever
reason sessionFactory() is not called (e.g. in test), DSF fails to give
us SQL connection.