MID-6319: added test+fix for NOT(nullable-prop-condition) treatment
Normally SQL does not return NULL for NOT x=y, but in our filter the NOT
has complement semantics and should return NULL values too.
This can be fixed by adding " AND x IS NOT NULL" inside the NOT just
after each condition inside.
Refactoring was required so that item filter processors can ask the
context, whether there is NOT filter above them.
During the refactoring SinglePathItemFilterProcessor emerged.
While currently all IFPs are extending it, PolyStringIFP will not.