MID-6319: fix of generic casting for JDK-8 around path.getClass() call
The base problem is that object.getClass() of Class<X> returns Class<?>.
When I used it as argument and later cast the result, it worked fine on
JDK 11 (and IDEA with Java 8 language level didn't indicate error).
But compiler failed with: jdk 8 incompatible types:
inferred type does not conform to equality constraint(s)

But if I cast that getClass call back to Class<X> everything was fine.