QuerydslUtils: removed expand(Path...), custom columns work differently
For custom columns in MID-6318, MID-6319 we don't need to expand
relational paths (entities) to all their columns, but we use structure
of Tuple [R, custom columns...], so we load the fixed part of the entity
in its bean and the custom columns as separate path expressions.
See SqlQueryContext#executeQuery.