Make ExpressionFactory thread-safe
In order to make ExpressionFactory thread-safe, more things had to be
done:
1) expression cache should be thread-safe (using ConcurrentHashMap),
2) keys should be really immutable, freezing or cloning them.

Point 2 is being resolved by freezing whatever can be frozen, and
cloning the rest. However, there are many things that are yet to be
frozen. So it's a work in progress.

Also fixing a bug that the cache didn't take expression profiles into
account.

Related to MID-6432.