New feature: Resource administrative operational state (aka resource maintenance)
Administrative operational state is set manually by midPoint users (e.g. administrator).
It signalizes whether resource is up and OPERATIONAL to receive provisioning requests or is down in MAINTENANCE and midPoint should not contact it during the provisioning.
This setting is useful when resource is e.g. undergoing planned maintenance to save computing time and error messages.
When under maintenance, operations are cached to the repository shadow and processed later when resource is back in OPERATIONAL mode.

EXPERIMENTAL FEATURE!