Evaluate normal mappings with secondary deltas
Originally, any delta on target item suppressed evaluation of normal
mappings for that target. This is understandable for primary deltas
but not so clear for secondary deltas. So, for secondary deltas
the behavior is changed now: normal mappings targeted to items that
have been changed previously (by secondary delta) are evaluated.

Related to MID-6406.