Clean up [Evaluated]Construction classes
Before we can implement "range" feature for construction
mappings we have to simplify/cleanup the class hierarchy
and the code a bit. This is the beginning. The relation
between ResourceObjectConstruction and OutboundConstruction
is not quite clear yet.

Related to MID-6228.

Also cleaned up working with "thisObject" variable. It is simply
computed from the assignment path on demand, not propagated throughout
the code in a messy way.

Fixed "source" variable when evaluating assignment target reference
filter.