Avoid re-processing of completed projections
The most important change is that we no longer clear
synchronizationPolicyDecision for completed projection contexts.
The primary motivation was to avoid loading deleted accounts but
the real consequences are much greater.

At some places (activation processing, assigned construction evaluation)
we added skipping when wave does not match and/or context is completed.

We also skip Projector.projectProjection if context is completed
(the wave check was already in place there).

When evaluating dependencies, we (maybe temporarily) relaxed
BROKEN/IGNORED check in wasProvisioned method. It was not functional
anyway, as synchronizationPolicyDecision was often cleared.

Related to MID-6228.