Fix activation computation and full shadow loading
Activation is now computed on each projector run but only for current
projections (projections with unknown wave are current by definition).
If existing synchronizationPolicyDecision is present, it is ignored.

Fixed loading full shadow for projections that are yet to be created.

Some tests were adapted in a minor way (regarding additional
full shadow load).

Related to MID-6228.