Add targetSelector to audit property definition
When the audit property definition was applied to a wrong type
of target object, errors sometimes occurred (sometimes not).
Although these were swallowed, they used to fill the logs and
disturb the admins.

Now, the targetSelector(s) can be used to specify objects to which
particular property definition should be applied.

Also, any problems during expression evaluation are now correctly
reported as partial errors in the operation result.

Related to MID-6592.