flexi-auth: support for simpler config of ignoredLocalPath for defaults
Previously the default sequences had to be declared explicitly,
otherwise ignoredLocalPath was ignored.
Now it's enough to just declare ignoredLocalPath and they will be used
for otherwise default setup.
Built-in ignored paths for actuator are not used then and must be
declared explicitly, but that's much easier than previous solution.