SqlRepositoryFactory is not "found" as a bean, but created by factory
RepositoryFactory is now Spring factory for RepositoryServiceFactory
which it returns based on the config.xml, the result is the same, but
it's much cleaner and allows future changes.
SqlRepositoryServiceImpl is then returned by SqlRepositoryFactory (which
is in facat SqlRepository_Service_Factory!) as before.
In line with this, SqaleRepositoryServiceFactory is not scanned
component either, it will be creatd by RepositoryFactory as needed.