Fix null-value "equals" filter in dummy resource
Dummy connector was fixed to work correctly with null-valued
equals filters. However, the solution is still too brittle,
mainly because of the ConnId EqualFilter implementation:
the exact state (null vs. empty collection) in the filter must
be equal to the one in the connector object for them to be matched.

Relates to MID-1460.