Make iterative tasks write op. execution records
A preliminary implementation of the recording of execution for complex
operations, i.e. operations covering the processing of the whole object
by a search-iterative task.

The functionality is provided by a triad of:
- OperationExecutionRecorderForClockwork (simple operations)
- OperationExecutionRecorderForTasks (complex operations within tasks)
- OperationExecutionWriter (the actual execution)

A work in progress.

!BEHAVIOR CHANGE! If we are not going to record anything, we won't do
ad-hoc cleanup as we did before. The cleanup should be done using
regular cleanup task.

Minor changes:
1. Introduced obligatory root task OID for RunnableTask objects,
so it is available for recording purposes at no cost.
2. Fixed isDelete method on projection context.
3. Fixed SystemConfigurationChangeDispatcherImpl: a thread-safety bug,
plus stopped listener update method returning (useless) boolean value.