Stop unlinking broken-being-deleted shadows
Originally the projCtx.isDelete method returned "false" for broken
contexts, even if the shadow deltas was DELETE. This was perhaps not
quite correct, so in 0bb36defb29448ba326713a8642439a64ba93160 we changed
this behavior. Now broken-being-deleted contexts have isDelete=true.

This means that such shadows started to be unlinked. It is maybe not
right (at least TestServiceAccounts.test104 thinks so). Therefore, here
we restore the original behavior of unlinking: broken-being-deleted
contexts are NOT unlinked.