Clean up InitializationState
The concepts related to processing initializable objects were unfinished
and misleading. Now they are still not 100% finished, but hopefully more
organized and logical.

In particular, the error and lifecycle states are considered separately.
Also the initialization flow is now in one method (instead of two).

Other changes: UcfChange identifiers are mutable again.