Allow emergency serialization of unsupported types
On rare occasions we allow prism property values to hold objects without
proper type name. A typical example is AuditEventRecord, processed by
auditing scripts. Such values are usually not serialized, so it causes
no harm.

But if they are to be serialized (e.g. when writing traces)
the serialization fails because the type is required.

So this commit adds an option to allow emergency serialization
that simply stores such values in their string form with type
set to xsd:string.

This resolves MID-6855.

Also:
- removed some unused code from HacksImpl, ExpressionUtil,
and SerializationOptions.