Fix entitlements handling during phantom renames
When a phantom rename is issued (e.g. DN is set to the same value
as it has had), the entitlements were added+removed, leading to their
loss. This is now fixed by checking if the rename is real.

This resolves MID-6770.

(cherry picked from commit 377aba6fd2899be715217efde9d35163bec71520)