Add message to operation execution records
New message field was added to OperationExecutionType. This is to
provide the necessary information also to complex records. (For simple
records such info has been in object delta list. But this is not
applicable to complex ones, because there are no natural deltas.)

So, from now on, please use the message from this place.

This resolves MID-6906.