Fix readonly option for get/search operations
Support for readonly option at the level of model service is tricky,
because returned objects are (almost) always modified e.g. because
of templates and security. So the fix consists of cloning such objects
(defeating the purpose of readonly option in the first place).

Also fixed readonly handling for provisioning.getObject call.

This resolves MID-6825.