Increase task execution intervals in some tests
TestReconTaskPartitioned and TestReconTaskMultinode use recurring
reconciliation tasks that originally started each 5 seconds. This
interval was so short that sometimes asserts that checked tasks'
state obtained newly started tasks instead of the ones that were
finished, and thous obviously failed.

The interval was increased to 10 seconds. Tests will run a little
longer but should be more stable.