Stop fetching all shadows during reconciliation
The 3rd part of reconciliation (shadows reconciliation) fetches all
shadows that were not processed in the 2nd part (resource
reconciliation) in order to find all of them that no longer have their
counterparts on the resource.

To find them it filters shadows on fullSynchronizationTimestamp.
Unfortunately, it compares that property with the task last start
timestamp, which is invalid for partitioned tasks. The result was that
all the shadows were (re-)processed during 3rd recon part.

Therefore now we have introduced lastReconciliationStartTimestamp task
extension property that is set when the real resource reconciliation
starts. And the comparison uses this property value.

This is a temporary solution, though. We assume that continuing task
cleanup/improvements to be done in 4.4 will provide more elegant options
for the implementation of this functionality.

This resolves MID-6930.