repo-sqlbase: introduced RepositoryException, above QueryException
When mapper supposed to be used for modification application is returned
it does not make any sense to throw QueryException, so something more
generic was needed.
In the end there is no change in the overal structure of try/catch and
the exception is mostly wrapped as SystemException as QEx before.