Fix a couple of task thread-safety issues
1) OperationResult is not a thread-safe structure, yet it was
exposed to multiple threads - mainly in getUpdatedOrClonedTaskObject
method.

2) Even backing task prism objects were exposed to multiple threads
in a special case: When a task with running LATs was fetched (with an
option to retrieve also subtasks), live instances of prism objects
were returned. If they were later processed - e.g. serialized in GUI
- concurrency-related exceptions could occur.

All of this should be now fixed, along with better naming for methods
that give access to backing task prism object.

A price to pay is that the operation result in returned objects
is not updated in these cases. But it should be acceptable.

This should resolve MID-6910.