Stop retrieving op. result from running tasks
When getting tasks from the repository, we tried to find their current
executions and updated the information from repo by the most up-to-date
values. Operation result was updated in this way as well.

Unfortunately, operation result is not thread-safe structure, so this
was inherently unreliable. Therefore, we no longer do this, at least
until OperationResult is not thread-safe.

This resolves MID-6215 and MID-6554.