Be more tolerant when non-ready task is started
When task in a state other than READY was started by Quartz, we used
to exit the execution, along with removing Quartz trigger. (Moreover
we did it very awkwardly, so an exception was propagated to Quartz.)

This caused issues in some tests, and (probably) would cause problems
in real life.

Therefore this commit relaxes this behavior. Now if a task is started
in unexpected (non-READY) state, the execution silently exits, without
removing the Quartz trigger. We hope that if the task is recurring,
it will be started again.