repo-sqale: all transformers merged into Q*Mapping classes
Mapping returned new transformer every time, but in the end transformer
was stateless too (most of the state is in other context objects).
Transformer support is now accessible statically to make this easier.
While some "abstract" mappers like QObjectMapping got bigger, it is much
easier to have just one concept instead of two.
Audit (non-sqale) part is not part of this commit.