Fix null values handling in DummyObject (approx.)
As the thread safety was ensured using ConcurrentHashMap
(and corresponding Set implementation), the null values can no
longer be stored in attributes. This is not a problem, as they
do not belong there anyway.

This commit filters out such values. Unfortunately, it does not do
that 100% consistently, so there are places (like sync delta creation)
where such values leak.