Fix various aspects of operation retrying
1. When provisioning - namely, the shadows module - updates a shadow
in repository (after e.g. successful ADD operation), it updates
in-memory information as well. This is a compensation for imprecise
code that replaced repo shadow by the resource object that was invoked
after retrying ADD operation.

2. We no longer retry modification or deletion operation if the resource
object does not exist (yet). This is to avoid useless failures.

3. When doing reconciliation, ContextLoader issues projection getObject
with forceRetry option only on the first occasion (projection wave 0).

4. Add forgotten clone when constructing pending operation delta.

This should resolve MID-7044.